// src/pages/Overview/views/Overview.enum.ts
export enum OverviewNetworkColumns {
    NumServices = 'Services',
    NumSites = 'Sites',
    NumDeployments = 'Deployments',
}

export enum OverviewLabels {
    Sites = 'Sites',
    Links = 'Links',
}

export enum OverviewLinksColumns {
    Source = 'From',
    Target = 'To',
    Cost = 'Cost',
    Mode = 'Mode',
}